### 2021-12-01 #3969 @whedon
[PRE REVIEW]: physical_validation: A Python package to assess the physical validity of molecular simulation results
**Submitting author:** @ptmerz (<a href="http://orcid.org/0000-0002-7045-8725">Pascal T. Merz</a>)
**Repository:** <a href="https://github.com/shirtsgroup/physical_validation" target ="_blank">https://github.com/shirtsgroup/physical_validation</a>
**Version:** v1.0
**Reviewers:** @tonigi, @geraldjwang and @zadehsey
