### 2021-12-07 #3981 @whedon
[REVIEW]: physical_validation: A Python package to assess the physical validity of molecular simulation results
**Submitting author:** @ptmerz (<a href="http://orcid.org/0000-0002-7045-8725">Pascal T. Merz</a>)
**Repository:** <a href="https://github.com/shirtsgroup/physical_validation" target ="_blank">https://github.com/shirtsgroup/physical_validation</a>
**Version:** v1.0.4

@tonigi & @geraldjwang and @zadehsey, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
