### 2021-12-14 #3981 @tonigi
The paper describes the physical_validation package, which performs a number of statistical tests grounded on the statistical mechanics of atomistic simulations. The need for such validation is clear and the package is useful. It does install and run easily, and it has state-of-the-art packaging (conda, pip, docs, tutorials and so on).

I only have minor comments on the structure of the paper.

1. It has a long "statement of need" section, a "prior work" section and... no "body"? It could make sense, for example, to make the second paragraph of the statement of need into a "Functionalities" paragraph, or similar.

2. As the authors note, the package was already described in the 2018 PLOS ONE paper. A cursory mention of the functionalities added since that release can help frame the current paper better.

3. GROMACS is mentioned and should be cited.

4. Regarding the software documentation, arguably a natural use case is to check on user-provided trajectory data. In addition to the GROMACS and LAMPPS parsers, wouldn't it make sense to use MDAnalysis to import trajectory data too? Ability to import data may also be mentioned in the manuscript.

Overall, good work.
