### 2021-12-14 #3981 @rkurchin
Just for reference: **[What should my paper contain?](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)** from JOSS docs

Worth noting that the two included sections are the only ones explicitly required (the general philosophy here is that the software is the primary product being evaluated, and the "paper" is really more there as a companion piece to provide context). That being said, there are no constraints on other allowable sections, and I agree it could be a bit clearer with the slight reorganization suggested.
