### 2021-12-15 c39a9ad
Address reviewer comments
* Split the "Statement of need" section
* Add GROMACS citation
