### 2021-12-16 e06f46d
Address reviewer comment: Add example of using MDAnalysis for trajectory
