+++ doc/simulation_data.rst
@@ -274,6 +274,46 @@ the equipartition check:
        seg = u.segments[i]
        molec_groups.append(np.array([seg.atoms[j].index for j in range(len(seg.atoms))]))
 
+Use :code:`MDAnalysis` to read position and velocity trajectory
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+:code:`MDAnalysis` also makes it easy to create `TrajectoryData` objects
+which require position and velocity trajectories as inputs. Given a
+`Universe` object which contains a trajectory, we can simply use a list
+comprehension to create a full trajectory in memory:
+::
+
+   import MDAnalysis as mda
+   import numpy as np
+   import physical_validation
+
+   u = mda.Universe('system.tpr', 'system.trr')
+   trajectory = physical_validation.data.TrajectoryData(
+       position=[frame.positions for frame in u.trajectory],
+       velocity=[frame.velocities for frame in u.trajectory])
+
+We can also use the atom selector to only feed part of the trajectory
+to the `physical_validation` tests.
+This is useful if we want to analyze the equipartition of parts of the
+system only (e.g. the solute) which can massively speed up the
+validation check. Note that we have to adapt the `SystemData` object
+accordingly to inform `physical_validation` that we are only analyzing
+part of the system.
+::
+
+   import MDAnalysis as mda
+   import numpy as np
+   import physical_validation
+
+   u = mda.Universe('system.tpr', 'system.trr')
+   protein = u.select_atoms('protein')
+   trajectory = physical_validation.data.TrajectoryData(
+       position=[protein.positions for _ in u.trajectory],
+       velocity=[protein.velocities for _ in u.trajectory])
+
+.. note:: `MDAnalysis` uses Å (ångström) as a length unit. Don't forget to
+          choose the `UnitData` accordingly!
+
 
 .. _simulationdata_details:
 
