### 2021-12-16 597d3a4
Address reviewer comment: Add example of using MDAnalysis for trajectory (#199)
