### 2021-12-16 e96e0a3
Bump version number to reflect documentation update
