+++ doc/conf.py
@@ -67,7 +67,7 @@ author = 'Pascal T. Merz, Michael R. Shirts'
 # built documents.
 #
 # The full version, including alpha/beta/rc tags.
-release = "1.0.1"
+release = "1.0.2"
 # The short X.Y version.
 version = ".".join(release.split(".")[:2])
 
