### 2021-12-16 #3981 @ptmerz
Thank you for your thoughtful review @tonigi!

> 1. It has a long "statement of need" section, a "prior work" section and... no "body"? It could make sense, for example, to make the second paragraph of the statement of need into a "Functionalities" paragraph, or similar.

I think that's a great suggestion! I added a "Functionality" header before the second paragraph of the "Statement of need" section, and it does indeed make the paper more clear.

> 2. As the authors note, the package was already described in the 2018 PLOS ONE paper. A cursory mention of the functionalities added since that release can help frame the current paper better.

This was the intention we had when writing second paragraph of the "Prior work" section. Do you have suggestions how to improve it?

> 3. GROMACS is mentioned and should be cited.

Agreed, I have added a citation.

> 4. Regarding the software documentation, arguably a natural use case is to check on user-provided trajectory data. In addition to the GROMACS and LAMPPS parsers, wouldn't it make sense to use MDAnalysis to import trajectory data too? Ability to import data may also be mentioned in the manuscript.

These are great suggestions. I added examples of how to import trajectory data from MDAnalysis at https://physical-validation.readthedocs.io/en/stable/simulation_data.html#use-mdanalysis-to-read-position-and-velocity-trajectory.

We are careful in advertising the ability to import data as an integral part of the package. The parsers are very useful, but we have learned that writing and maintaining such parsers is extremely difficult. The large array of functionality of the molecular simulation packages makes it very hard to cover all (edge) cases, and simulation input and output files are often rather unstable formats which tend to change between versions as functionality is added or removed. We decided to keep shipping the parsers since they are useful to many users and correct to the best of our knowledge. We have, however, focussed the documentation on providing the input to `physical_validation` by other means (e.g. with the help of Python APIs of the software packages or other libraries such as `MDAnalysis`), which we think is the more future-proof way to interface with the package.

Again, I really appreciate your time and your thoughtful review!
