### 2021-12-16 #3981 @tonigi
Thanks for the response and the changes. The fact that the new features are under the heading "Prior work" is slightly confusing, but ok I guess.  I have no further comments :)
