### 2021-12-16 #3981 @ptmerz
I see what you're saying now, thanks for clarifying. Maybe renaming the section to "Relation to prior work" would make it less confusing?
