### 2021-12-16 #3981 @tonigi
Sounds a reasonable solution.
