### 2021-12-22 #3981 @zadehsey
I really enjoyed the idea behind the work and believe many would benefit from it. Going over the documentation and example files, I had a number of minor suggestions:

Integrator_validation

- The cutoff radius of the LJ potential is given as “1 nm”. I believe that it will be more generalizable if the values are non-dimensionalized with respect to the parameters of the interatomic potential. The same applies to the timestep of the simulation as well.
- The plot axis needs to be fixed. Potentially using a scientific format. Also, it will be easier to read the plot if a marker is added where the actual data point is. Especially when the behaviour is non-monotonic

Kinetic_energy_distribution

- I think that it will be easier to define variables for number of atoms per molecule, degree of constraints, etc. so the user will change those and the description of the variable is also more clear
- Visually, it may be useful to add a vertical line showing the mean values obtained from the trajectory and the analytical solution
