### 2021-12-22 791543e
Address JOSS reviewer comments
* Use variables in kinetic energy deistribution example notebook,
  creation of system data structure
* Add symbol at integrator plot simulation points
* Fix unrelated typo in integrator notebook
