### 2021-12-22 #3981 @ptmerz
Thank you for your comments @zadehsey!

> Integrator_validation
>
> * The cutoff radius of the LJ potential is given as “1 nm”. I believe that it will be more generalizable if the values are non-dimensionalized with respect to the parameters of the interatomic potential. The same applies to the timestep of the simulation as well.

I agree that from a theoretical perspective, using simulation parameters which are depending on interaction or other system parameters makes a lot of sense. In the scope of the physical validation analysis tool, I do however think that it is more useful to have examples that use the type of parameters that users are required to set. In many biomolecular simulation packages, setting absolute values is the norm. See for example GROMACS ([parameter dt](https://manual.gromacs.org/documentation/current/user-guide/mdp-options.html#mdp-dt) in ps, [parameter rvdw](https://manual.gromacs.org/documentation/current/user-guide/mdp-options.html#mdp-rvdw) in nm),  OpenMM([setStepSize](https://openmmtools.readthedocs.io/en/0.17.0/api/generated/openmmtools.integrators.ThermostatedIntegrator.html#openmmtools.integrators.ThermostatedIntegrator.setStepSize) in ps, [setCutoffDistance function](http://docs.openmm.org/7.2.0/api-c++/generated/OpenMM.NonbondedForce.html#_CPPv2N6OpenMM14NonbondedForce17setCutoffDistanceEd) in nm), LAMMPS ([timestep command](https://docs.lammps.org/timestep.html) in time units, [lj/cut command](https://docs.lammps.org/pair_lj.html) sets LJ cutoff in distance units [which is not depending on the LJ parameters except for units lj]). I would therefore prefer to leave the examples in absolute units, if you don't have strong objections!

> * The plot axis needs to be fixed. Potentially using a scientific format. Also, it will be easier to read the plot if a marker is added where the actual data point is. Especially when the behaviour is non-monotonic

Fixing the plot axis is a non-trivial problem. Note that the tool is plotting plots as a visual aid, but is not aiming at producing publication-ready figures. The latter is an almost impossible task given that there is no way to know the values of the integrator convergence before performing the analysis. I fully agree that in the example, it would be nice to have the plots all sharing the same fixed axis. But they come from three independent analysis runs, and I don't see an obvious way how to handle this in general.

I think that having a plot marker is a great thought. I implemented this suggestion in https://github.com/shirtsgroup/physical_validation/pull/201, and the updated documentation looks like this https://physical-validation.readthedocs.io/en/latest/examples/integrator_validation.html. The changes are merged, and I will push a new stable version containing this patch once we conclude this review!

> Kinetic_energy_distribution
>
> * I think that it will be easier to define variables for number of atoms per molecule, degree of constraints, etc. so the user will change those and the description of the variable is also more clear

I changes the notebook to have these in variables. See https://physical-validation.readthedocs.io/en/latest/examples/kinetic_energy_distribution.html.

> * Visually, it may be useful to add a vertical line showing the mean values obtained from the trajectory and the analytical solution

I agree very much with your thought - we actually had that implemented for a while 😃 We found, however, that in any real-world use, the mean values of total kinetic energy estimates are almost always extremely close. (NB: This is actually one of the reason the *distribution* test is so useful - the mean might be correct, but if the distribution is not followed, the simulation might still be physically invalid!) The vertical lines added more confusion than clarity, so we decided to remove them.
