### 2021-12-23 #3981 @zadehsey
Thank you for your response @ptmerz.
Regarding the first point, I do not have a strong objection and understand why you would want to have the values presented in the current format. However, as the example used is specifically LJ, I believe that at least mentioning the length-scale and energy-scale there as a reference would be helpful for those who prefer to think in non-dimensional terms
