### 2021-12-23 #3981 @ptmerz
That is a good point @zadehsey. Adding some information on the LJ parameters used and setting the cutoff in relation to that adds value to the example. I have done that in https://github.com/shirtsgroup/physical_validation/pull/202 and it is now reflected in the documentation on https://physical-validation.readthedocs.io/en/latest/examples/integrator_validation.html.

Thanks again for your suggestion!
