### 2021-12-30 #3981 @geraldjwang
@ptmerz and team: I wanted to, above all else, congratulate you on a thoughtful (and delightful (and sorely needed)) package, which has really been an end-of-a year treat to look through! (Seriously!) I appreciate the thoughtful iterations w.r.t. points raised by @zadehsey and @tonigi. Our community needs this kind of work. As far as the formal review process, I have nothing further to request: Everything on the checklist is clearly there, and this is overall a wonderful (and, again, sorely needed) step in the direction of more automated and more physics-grounded verification of MD results.

I have several ideas for additional tests that could be quite informative to develop within this framework, but I'll keep that discussion separate from the review process since these ideas are immaterial to the suitability of this work as is (this work is, in my view, wonderfully suitable for JOSS in its current form).

I also wish to add (although this is also immaterial to the formal review) that I will be absolutely delighted to point students towards this package in my molecular simulation course when I am teaching it again next year. (To be clear, they will still have to verify that, e.g., their simulation results satisfy Equipartition Theorem via their own elbow grease, but it's wonderful to be able to then show them that tools such as yours exist, and are actively being developed by thoughtful people!)
