### 2022-01-03 d50b0ac
Implement editorial suggestions
