### 2022-01-03 #3981 @ptmerz
Thank you for your time and your thoughtful review @geraldjwang. It am very happy to hear that you find our work useful - reading your comment was indeed a nice way to start the new year :)

As for your suggestions for additional tests, please do not hesitate to get in touch either via the GitHub issue tracker, or directly via e-mail. We would love to add new functionality that is useful to the community, so we definitely welcome ideas and / or contributions.
