### 2022-01-03 7bb3ccd
Joss paper (#203)
* Initial paper commit

* Updated author list

* suggestions/additions for JOSS paper (#169)

* Fix references

* Update paper.md (#195)

Adding some funding acknowledgements.

* Address JOSS reviewer comments (#198)

Addresses reviewer comments of first pre-review by fixing references
and explaining relation to prior work.

* Fix one word (paper -> package).

* Added a word

* Address reviewer comments

* Split the "Statement of need" section
* Add GROMACS citation

* Address reviewer comment by renaming "Prior work" to "Relation to prior work" for added clarity

* Implement
[…]
