+++ joss/paper.bib
@@ -1,0 +1,38 @@
+@article{Merz:2018,
+    doi = {10.1371/journal.pone.0202764},
+    author = {{Merz}, Pascal T. and {Shirts}, Michael R.},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    title = {Testing for physical validity in molecular simulations},
+    year = {2018},
+    month = {09},
+    volume = {13},
+    url = {https://doi.org/10.1371/journal.pone.0202764},
+    pages = {1-22},
+    number = {9},
+}
+@article{Shirts:2013,
+    doi = {10.1021/ct300688p},
+    author = {{Shirts}, Michael R.},
+    journal = {Journal of Chemical Theory and Computation},
+    title = {Simple Quantitative Tests to Validate Sampling from Thermodynamic Ensembles},
+    year = {2013},
+    volume = {9},
+    url = {https://doi.org/10.1021/ct300688p},
+    pages = {909-926},
+    number = {2},
+}
+@article{Abraham:2015,
+title = {GROMACS: High performance molecular simulations through multi-level parallelism from laptops to supercomputers},
+journal = {SoftwareX},
+volume = {1-2},
+pages = {19-25},
+year = {2015},
+issn = {2352-7110},
+doi = {https://doi.org/10.1016/j.softx.2015.06.001},
+url = {https://www.sciencedirect.com/science/article/pii/S2352711015000059},
+author = {Mark James Abraham and Teemu Murtola and Roland Schulz and Szilárd Páll and Jeremy C. Smith and Berk Hess and Erik Lindahl},
+keywords = {Molecular dynamics, GPU, SIMD, Free energy},
+abstract = {GROMACS is one of the most widely used open-source and free software codes in chemistry, used primarily for dynamical simulations of biomolecules. It provides a rich set of calculation types, preparation and analysis tools. Several advanced techniques for free-energy calculations are supported. In version 5, it reaches new performance heights, through several new and enhanced parallelization algorithms. These work on every level; SIMD registers inside cores, multithreading, heterogeneous CPU–GPU acceleration, state-of-the-art 3D domain decomposition, and ensemble-level parallelization through built-in replica exchange and the separate Copernicus framework. The latest best-in-class compressed trajectory storage format is supported.}
+}
+
