### 2022-01-03 8adf099
Bump version after JOSS review (#204)
