+++ doc/conf.py
@@ -59,7 +59,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'physical_validation'
-copyright = '2017-2021, Pascal T. Merz, Michael R. Shirts'
+copyright = '2017-2022, Pascal T. Merz, Michael R. Shirts'
 author = 'Pascal T. Merz, Michael R. Shirts'
 
 # The version info for the project you're documenting, acts as replacement for
@@ -67,7 +67,7 @@ author = 'Pascal T. Merz, Michael R. Shirts'
 # built documents.
 #
 # The full version, including alpha/beta/rc tags.
-release = "1.0.2"
+release = "1.0.3"
 # The short X.Y version.
 version = ".".join(release.split(".")[:2])
 
