+++ README.md
@@ -1,6 +1,3 @@
-Physical validation reference
-=============================
-
 [![GitHub Actions Test Status](https://github.com/shirtsgroup/physical_validation/actions/workflows/continous_integration.yaml/badge.svg)](https://github.com/shirtsgroup/physical_validation/actions/workflows/continous_integration.yaml)
 [![GitHub Actions Lint Status](https://github.com/shirtsgroup/physical_validation/actions/workflows/lint.yaml/badge.svg)](https://github.com/shirtsgroup/physical_validation/actions/workflows/lint.yaml)
 [![Documentation Status](https://readthedocs.org/projects/physical-validation/badge/?version=latest)](https://physical-validation.readthedocs.io/en/latest/?badge=latest)
@@ -10,20 +7,20 @@ Physical validation reference
 [![DOI](https://zenodo.org/badge/90674371.svg)](https://zenodo.org/badge/latestdoi/90674371)  
 [![MIT license](https://img.shields.io/badge/License-MIT-blue.svg)](https://lbesson.mit-license.org/)
 
+physical_validation: A Python package to assess the physical validity of molecular simulation results
+=====================================================================================================
 
-
-`physical_validation` is a package aimed at testing results obtained
-by molecular dynamics simulations for their physical validity.
+`physical_validation` is a package testing results obtained by molecular simulations for their physical validity.
 
 Please cite 
 > Merz PT, Shirts MR (2018) Testing for physical validity in molecular simulations. PLoS ONE 13(9): e0202764. https://doi.org/10.1371/journal.pone.0202764
 
+`physical_validation` incorporates the functionality of [checkensemble](https://github.com/shirtsgroup/checkensemble).
+
+This software is developed in the Shirts group at University of Colorado in Boulder.
+
+Documentation
+-------------
 Please check
 [https://physical-validation.readthedocs.io](http://physical-validation.readthedocs.io)
 for the full reference.
-
-`physical_validation` incorporates the functionality of
-[checkensemble](https://github.com/shirtsgroup/checkensemble).
-
-This software is developed in the Shirts group at University of 
-Colorado in Boulder.
