### 2022-01-03 2aa3be9
Fix DOI metadata & bump version
