### 2022-01-03 #3981 @ptmerz
Thank you for your editorial suggestions @rkurchin, they are much appreciated!

All of the reviewers' suggestions have been merged. I have also included your editorial suggestions in the manuscript, and merged the manuscript into the main branch.

I created a new version including all the latest changes:
* Version tag: v1.0.4.
* DOI: 10.5281/zenodo.5815657
* Zenodo link: https://zenodo.org/record/5815657#.YdM1dy-B0oY
* GitHub Release https://github.com/shirtsgroup/physical_validation/releases/tag/v1.0.4
* PyPI: https://pypi.org/project/physical-validation/1.0.4/
* Conda-forge will be updated in a few hours when the bots will have caught up :)
