### 2022-01-03 a4d5141
Fix DOI in JOSS paper reference
