+++ joss/paper.bib
@@ -29,7 +29,7 @@ volume = {1-2},
 pages = {19-25},
 year = {2015},
 issn = {2352-7110},
-doi = {https://doi.org/10.1016/j.softx.2015.06.001},
+doi = {10.1016/j.softx.2015.06.001},
 url = {https://www.sciencedirect.com/science/article/pii/S2352711015000059},
 author = {Mark James Abraham and Teemu Murtola and Roland Schulz and Szilárd Páll and Jeremy C. Smith and Berk Hess and Erik Lindahl},
 keywords = {Molecular dynamics, GPU, SIMD, Free energy},
