### 2022-01-03 #3981 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
