### 2022-01-05 #3981 @kthyng
Zenodo archive and version check are good!

Paper:
* citation in first paragraph shouldn't be parenthetical — remove the [] around citation. Some later citations have this problem too.
