### 2022-01-06 8dd4538
Fix JOSS citations
All citations were parenthetical, fixed to use paranthesis only where
it makes sense.
