### 2022-01-06 #3981 @ptmerz
Thanks @kthyng, it seems we just used the parenthetical version throughout without taking into account the context. I went over all citations and used the version (with or without parenthesis) which fits the context better.
